clean up usare-role controller

+++ src/Ushahidi/Modules/V5/Http/Controllers/UserController.php
@@ -13,13 +13,9 @@ use Ushahidi\Modules\V5\Actions\User\Queries\FetchUserQuery;
 use Ushahidi\Modules\V5\Actions\User\Commands\CreateUserCommand;
 use Ushahidi\Modules\V5\Actions\User\Commands\DeleteUserCommand;
 use Ushahidi\Modules\V5\Actions\User\Commands\UpdateUserCommand;
-use Ushahidi\Core\Exception\AuthorizerException;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Contracts\Auth\Access\Gate;
 use Ushahidi\Core\Entity\User as UserEntity;
 use Ushahidi\Modules\V5\DTO\UserSearchFields;
-//use Ushahidi\Core\Tool\Hasher\Password as PasswordHash;
 use Ushahidi\Modules\V5\Requests\UserRequest;
 use Illuminate\Support\Facades\Log;
 use Ushahidi\Core\Exception\NotFoundException;
@@ -45,7 +41,7 @@ class UserController extends V5Controller
     public function show(int $id)
     {
         $user = $this->queryBus->handle(new FetchUserByIdQuery($id));
-        $this->authorizeForCurrentUserForUser('show', $user);
+        $this->authorize('show', $user);
         return new UserResource($user);
     } //end show()
 
@@ -59,12 +55,11 @@ class UserController extends V5Controller
      */
     public function showMe()
     {
-        $id = $this->getGenericUserForUser()->id;
+        $id = AUTH::id();
         if (!$id) {
             throw new NotFoundException('User not found');
         }
         $user = $this->queryBus->handle(new FetchUserByIdQuery($id));
-        $this->authorizeForCurrentUserForUser('show', $user);
         return new UserResource($user);
     } //end show()
 
@@ -76,7 +71,8 @@ class UserController extends V5Controller
      */
     public function index(Request $request)
     {
-        $this->authorizeForCurrentUserForUser('index', User::class);
+        $this->authorize('index', new User());
+
         $resourceCollection = new UserCollection(
             $this->queryBus->handle(
                 new FetchUserQuery(
@@ -96,12 +92,13 @@ class UserController extends V5Controller
      * Create new User.
      *
      * @param UserRequest $request
-     * @return \Illuminate\Http\JsonResponse|CategoryResource
+     * @return \Illuminate\Http\JsonResponse|UserResource
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(UserRequest $request)
     {
-        $this->authorizeForCurrentUserForUser('store', User::class);
+        $this->authorize('store', new User());
+
         $command = new CreateUserCommand(UserEntity::buildEntity($request->input()));
         $this->commandBus->handle($command);
         return new UserResource(
@@ -120,7 +117,8 @@ class UserController extends V5Controller
     public function update(UserRequest $request, int $id)
     {
         $user = $this->queryBus->handle(new FetchUserByIdQuery($id));
-        $this->authorizeForCurrentUserForUser('update', $user);
+        $this->authorize('update', $user);
+
         $this->commandBus->handle(
             new UpdateUserCommand($id, UserEntity::buildEntity($request->input(), 'update', $user->toArray()))
         );
@@ -138,15 +136,12 @@ class UserController extends V5Controller
      */
     public function updateMe(UserRequest $request)
     {
-        $id = AUTH::id();//$this->getGenericUserForUser()->id;
+        $id = AUTH::id();
         $user = $this->queryBus->handle(new FetchUserByIdQuery($id));
-        $this->authorizeForCurrentUserForUser('update', $user);
         $this->commandBus->handle(
             new UpdateUserCommand($id, $this->buildUserEntity("update", $request, $user))
         );
-        return new UserResource(
-            $this->queryBus->handle(new FetchUserByIdQuery($id))
-        );
+        return $this->showMe();
     } //end update()
 
 
@@ -160,24 +155,8 @@ class UserController extends V5Controller
     public function delete(int $id)
     {
         $user = $this->queryBus->handle(new FetchUserByIdQuery($id));
-        $this->authorizeForCurrentUserForUser('delete', $user);
+        $this->authorize('delete', $user);
         $this->commandBus->handle(new DeleteUserCommand($id));
         return $this->deleteResponse($id);
     } //end store()
-
-
-
-    // To Do : Replace with authorizeForCurrentUser after merge
-    private function authorizeForCurrentUserForUser($ability, $arguments = [])
-    {
-        $gUser = $this->getGenericUserForUser();
-
-        list($ability, $arguments) = $this->parseAbilityAndArguments($ability, $arguments);
-        return app(Gate::class)->forUser($gUser)->authorize($ability, $arguments);
-    }
-
-    private function getGenericUserForUser()
-    {
-        return Auth::guard()->user();
-    }
 } //end class
